#914 - Modularize project import/export
- Relocated rest of the core exporters to the api-dao module
